+++ src/components/victory-chart/helper-methods.js
@@ -1,15 +1,139 @@
+/* eslint-disable func-style */
+/* eslint-disable no-use-before-define */
+
 import Axis from "../../helpers/axis";
 import Wrapper from "../../helpers/wrapper";
 import React from "react";
-import { Collection, Log, Helpers } from "victory-core";
-import { isFunction } from "lodash";
+import { Log, Helpers, Scale } from "victory-core";
+import { defaults, assign } from "lodash";
 
-const getDataComponents = (childComponents) => {
-  const iteratee = (child) => {
-    return child.type && isFunction(child.type.getData) ? child : null;
+function getAxisProps(child, props, calculatedProps) {
+  const { domain, scale, originSign, stringMap, categories, horizontal } = calculatedProps;
+  const childProps = child.props || {};
+  const axis = child.type.getAxis(childProps);
+  const currentAxis = Axis.getCurrentAxis(axis, horizontal);
+  const otherAxis = axis === "x" ? "y" : "x";
+  const axisOffset = getAxisOffset(props, calculatedProps);
+  const offsetY = axis === "y" ? undefined : axisOffset.y;
+  const offsetX = axis === "x" ? undefined : axisOffset.x;
+  const crossAxis = childProps.crossAxis === false ? false : true;
+  const orientation = Axis.getOrientation(child, axis, originSign[otherAxis]);
+  return {
+    stringMap: stringMap[currentAxis],
+    categories: categories[currentAxis],
+    startAngle: props.startAngle,
+    endAngle: props.endAngle,
+    innerRadius: props.innerRadius,
+    domain,
+    scale,
+    offsetY: childProps.offsetY !== undefined ? childProps.offsetY : offsetY,
+    offsetX: childProps.offsetX !== undefined ? childProps.offsetX : offsetX,
+    crossAxis,
+    orientation
   };
-  return Helpers.reduceChildren(childComponents.slice(0), iteratee);
-};
+}
+
+function getChildProps(child, props, calculatedProps) {
+  const axisChild = Axis.findAxisComponents([child]);
+  if (axisChild.length > 0) {
+    return getAxisProps(axisChild[0], props, calculatedProps);
+  }
+  const { categories, domain, range, scale } = calculatedProps;
+  return { categories, domain, range, scale };
+}
+
+function getStyles(props) {
+  const styleProps = props.style && props.style.parent;
+  return {
+    parent: defaults(
+      {},
+      styleProps,
+      {
+        height: "100%",
+        width: "100%",
+        userSelect: "none"
+      }
+  ) };
+}
+
+function getCalculatedProps(props, childComponents) {
+  const style = getStyles(props);
+  const horizontal = childComponents.some((component) => {
+    return component.props && component.props.horizontal;
+  });
+  // TODO: check
+  const categories = {
+    x: Wrapper.getCategories(props, "x", childComponents),
+    y: Wrapper.getCategories(props, "y", childComponents)
+  };
+  const stringMap = {
+    x: createStringMap(props, "x", childComponents),
+    y: createStringMap(props, "y", childComponents)
+  };
+  const axisComponents = {
+    x: Axis.getAxisComponent(childComponents, "x"),
+    y: Axis.getAxisComponent(childComponents, "y")
+  };
+  const domain = {
+    x: getDomain(assign({}, props, { categories }), "x", childComponents),
+    y: getDomain(assign({}, props, { categories }), "y", childComponents)
+  };
+  const range = {
+    x: Helpers.getRange(props, "x"),
+    y: Helpers.getRange(props, "y")
+  };
+  const baseScale = {
+    x: Scale.getScaleFromProps(props, "x") ||
+      axisComponents.x && axisComponents.x.type.getScale(axisComponents.x.props) ||
+      Scale.getDefaultScale(),
+    y: Scale.getScaleFromProps(props, "y") ||
+      axisComponents.y && axisComponents.y.type.getScale(axisComponents.y.props) ||
+      Scale.getDefaultScale()
+  };
+  const scale = {
+    x: baseScale.x.domain(domain.x).range(range.x),
+    y: baseScale.y.domain(domain.y).range(range.y)
+  };
+
+  const origin = props.polar ? Helpers.getPolarOrigin(props) : Axis.getOrigin(domain);
+
+  const originSign = {
+    x: Axis.getOriginSign(origin.x, domain.x),
+    y: Axis.getOriginSign(origin.y, domain.y)
+  };
+
+  const defaultDomainPadding = getDefaultDomainPadding(childComponents, horizontal);
+
+  const padding = Helpers.getPadding(props);
+
+  return {
+    axisComponents, categories, domain, range, horizontal, scale, stringMap,
+    style, origin, originSign, defaultDomainPadding, padding
+  };
+}
+
+function getChildren(props, childComponents, calculatedProps) {
+  childComponents = childComponents || getChildComponents(props);
+  calculatedProps = calculatedProps || getCalculatedProps(props, childComponents);
+  const baseStyle = calculatedProps.style.parent;
+  const { height, polar, theme, width } = props;
+  const { origin } = calculatedProps;
+  return childComponents.map((child, index) => {
+    const style = Array.isArray(child.props.style) ?
+      child.props.style :
+      defaults({}, child.props.style, { parent: baseStyle });
+    const childProps = getChildProps(child, props, calculatedProps);
+    const newProps = defaults({
+      height, polar, theme, width, style,
+      origin: polar ? origin : undefined,
+      padding: calculatedProps.padding,
+      key: index,
+      standalone: false
+    }, childProps);
+
+    return React.cloneElement(child, newProps);
+  });
+}
 
 const getChildComponents = (props, defaultAxes) => {
   const childComponents = React.Children.toArray(props.children);
@@ -25,17 +149,23 @@ const getChildComponents = (props, defaultAxes) => {
   if (axisComponents.dependent.length === 0 && axisComponents.independent.length === 0) {
     return childComponents.concat([defaultAxes.independent, defaultAxes.dependent]);
   }
-  if (axisComponents.independent.length > 1) {
-    const msg = "Only one independent VictoryAxis component is allowed when " +
-      "using the VictoryChart wrapper. Only the first axis will be used. Please compose " +
-      "multi-axis charts manually";
-    Log.warn(msg);
-    const dataComponents = getDataComponents(childComponents);
-    return Collection.removeUndefined(
-      dataComponents.concat([...axisComponents.dependent, axisComponents.independent[0]])
-    );
-  }
-  return childComponents;
+  let axisCount = 0;
+  return childComponents.filter((child) => {
+    const role = child.type && child.type.role;
+    const childProps = child.props || {};
+    if (role !== "axis" || !childProps.dependentAxis) {
+      return true;
+    } else if (axisCount < 1) {
+      axisCount++;
+      return true;
+    } else {
+      const msg = "Only one independent VictoryAxis component is allowed when " +
+        "using the VictoryChart wrapper. Only the first axis will be used. Please compose " +
+        "multi-axis charts manually";
+      Log.warn(msg);
+      return false;
+    }
+  });
 };
 
 const getDefaultDomainPadding = (childComponents, horizontal) => {
@@ -104,10 +234,8 @@ const createStringMap = (props, axis, childComponents) => {
     }, {});
 };
 
-export { getDomain,
-  getAxisOffset,
-  getDataComponents,
-  getChildComponents,
-  getDefaultDomainPadding,
-  createStringMap
+export {
+  getChildren,
+  getCalculatedProps,
+  getChildComponents
 };
