+++ src/components/victory-stack/helper-methods.js
@@ -1,6 +1,6 @@
 /* eslint-disable func-style */
 /* eslint-disable no-use-before-define */
-import { assign } from "lodash";
+import { assign, keys, orderBy } from "lodash";
 import React from "react";
 import { Helpers, Scale } from "victory-core";
 import Wrapper from "../../helpers/wrapper";
@@ -10,6 +10,93 @@ const fallbackProps = {
   height: 300,
   padding: 50
 };
+
+// Assumes data in `datasets` is sorted by `Data.getData`.
+function fillData(props, datasets) {
+  const { fillInMissingData } = props;
+  const xMap = datasets.reduce((prev, dataset) => {
+    dataset.forEach((datum) => {
+      prev[datum._x instanceof Date ? datum._x.getTime() : datum._x] = true;
+    });
+    return prev;
+  }, {});
+  const xKeys = keys(xMap).map((k) => +k);
+  const xArr = orderBy(xKeys);
+
+  return datasets.map((dataset) => {
+    let indexOffset = 0;
+    const isDate = dataset[0] && dataset[0]._x instanceof Date;
+    const filledInData = xArr.map((x, index) => {
+      x = +x;
+      const datum = dataset[index - indexOffset];
+
+      if (datum) {
+        const x1 = isDate ? datum._x.getTime() : datum._x;
+        if (x1 === x) {
+          return datum;
+        } else {
+          indexOffset++;
+          const y = fillInMissingData ? 0 : null;
+          x = isDate ? new Date(x) : x;
+          return { x, y, _x: x, _y: y };
+        }
+      } else {
+        const y = fillInMissingData ? 0 : null;
+        x = isDate ? new Date(x) : x;
+        return { x, y, _x: x, _y: y };
+      }
+    });
+
+    return filledInData;
+  });
+}
+
+function getY0(datum, index, datasets) {
+  if (datum.y0) {
+    return datum.y0;
+  }
+  const y = datum._y;
+  const previousDatasets = datasets.slice(0, index);
+  const previousPoints = previousDatasets.reduce((prev, dataset) => {
+    return prev.concat(dataset
+      .filter((previousDatum) => datum._x instanceof Date
+        ? previousDatum._x.getTime() === datum._x.getTime()
+        : previousDatum._x === datum._x)
+      .map((previousDatum) => previousDatum._y || 0)
+    );
+  }, []);
+  const y0 = previousPoints.length && previousPoints.reduce((memo, value) => {
+    const sameSign = (y < 0 && value < 0) || (y >= 0 && value >= 0);
+    return sameSign ? +value + memo : memo;
+  }, 0);
+  return previousPoints.some((point) => point instanceof Date) ? new Date(y0) : y0;
+}
+
+/* eslint-disable no-nested-ternary */
+function addLayoutData(props, datasets, index) {
+  const xOffset = props.xOffset || 0;
+  return datasets[index].map((datum) => {
+    const yOffset = getY0(datum, index, datasets) || 0;
+    return assign({}, datum, {
+      _y0: !(datum._y instanceof Date) ? yOffset : (
+        yOffset ? new Date(yOffset) : datum._y
+      ),
+      _y1: datum._y === null ? null : (
+        datum._y instanceof Date ? new Date(+datum._y + +yOffset) : datum._y + yOffset
+      ),
+      _x1: datum._x === null ? null : (
+        datum._x instanceof Date ? new Date(+datum._x + +xOffset) : datum._x + xOffset
+      )
+    });
+  });
+}
+/* eslint-enable no-nested-ternary */
+
+function stackData(props) {
+  const dataFromChildren = Wrapper.getDataFromChildren(props);
+  const datasets = fillData(props, dataFromChildren);
+  return datasets.map((d, i) => addLayoutData(props, datasets, i));
+}
 
 function getCalculatedProps(props, childComponents) {
   childComponents = childComponents || React.Children.toArray(props.children);
@@ -22,11 +109,13 @@ function getCalculatedProps(props, childComponents) {
     x: Wrapper.getCategories(props, "x"),
     y: Wrapper.getCategories(props, "y")
   };
-  const dataFromChildren = Wrapper.getDataFromChildren(props);
-  const datasets = Wrapper.fillInMissingData(props, dataFromChildren);
+  const datasets = stackData(props);
+  const children = childComponents.map((c, i) => {
+    return React.cloneElement(c, { data: datasets[i] });
+  });
   const domain = {
-    x: Wrapper.getStackedDomain(assign({}, props, { categories }), "x", datasets),
-    y: Wrapper.getStackedDomain(assign({}, props, { categories }), "y", datasets)
+    x: Wrapper.getDomain(assign({}, props, { categories }), "x", children),
+    y: Wrapper.getDomain(assign({}, props, { categories }), "y", children)
   };
   const range = {
     x: Helpers.getRange(props, "x"),
@@ -87,7 +176,7 @@ function getChildren(props, childComponents, calculatedProps) {
   const childProps = getChildProps(props, calculatedProps);
 
   return childComponents.map((child, index) => {
-    const data = Wrapper.addLayoutData(props, datasets, index);
+    const data = datasets[index];
     const style = Wrapper.getChildStyle(child, index, calculatedProps);
     const labels = props.labels ? getLabels(props, datasets, index) : child.props.labels;
